feat(fetchBooks):creating books slice and fetch
