+++ src/store/books/bookSlice.js
@@ -1,0 +1,54 @@
+import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
+import data from "../../service/fakeData.json";
+const initialState = {
+  entities: [],
+  status: "idle",
+};
+
+// The function below is called a thunk and allows us to perform async logic. It
+// can be dispatched like a regular action: `dispatch(incrementAsync(10))`. This
+// will call the thunk with the `dispatch` function as the first argument. Async
+// code can then be executed and other actions can be dispatched. Thunks are
+// typically used to make async requests.
+export const fetchBooks = createAsyncThunk("books/fetchBooks", async () => {
+  const response = await data;
+  // The value we return becomes the `fulfilled` action payload
+  return response;
+});
+
+export const bookSlice = createSlice({
+  name: "books",
+  initialState,
+  // The `reducers` field lets us define reducers and generate associated actions
+  reducers: {},
+  // The `extraReducers` field lets the slice handle actions defined elsewhere,
+  // including actions generated by createAsyncThunk or in other slices.
+  extraReducers: (builder) => {
+    builder
+      .addCase(fetchBooks.pending, (state) => {
+        state.status = "loading";
+      })
+      .addCase(fetchBooks.fulfilled, (state, action) => {
+        state.status = "idle";
+        state.entities = action.payload;
+      });
+  },
+});
+
+// export const { increment, decrement, incrementByAmount } = counterSlice.actions;
+
+// The function below is called a selector and allows us to select a value from
+// the state. Selectors can also be defined inline where they're used instead of
+// in the slice file. For example: `useSelector((state: RootState) => state.counter.value)`
+// export const selectCount = (state) => state.counter.value;
+
+// We can also write thunks by hand, which may contain both sync and async logic.
+// Here's an example of conditionally dispatching actions based on current state.
+// export const incrementIfOdd = (amount) => (dispatch, getState) => {
+//   const currentValue = selectCount(getState());
+//   if (currentValue % 2 === 1) {
+//     dispatch(incrementByAmount(amount));
+//   }
+// };
+
+export default bookSlice.reducer;
