+++ src/store/index.js
@@ -1,8 +1,8 @@
 import { configureStore } from "@reduxjs/toolkit";
-import counterReducer from "./counter/counterSlice";
+import booksReducer from "./books/bookSlice";
 
 export const store = configureStore({
   reducer: {
-    counter: counterReducer,
+    books: booksReducer,
   },
 });
